^/sandbox/online: various changes to online-decoding setup for efficiency.  Will make further changes.
git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/online@3669 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/base/kaldi-math.h
@@ -101,8 +101,8 @@ inline float RandUniform() {  // random intended to be strictly between 0 and 1.
 }
 
 inline float RandGauss() {
-  return static_cast<float>(sqrt (-2 * std::log(RandUniform()))
-                            * cos(2*M_PI*RandUniform()));
+  return static_cast<float>(sqrtf (-2 * logf(RandUniform()))
+                            * cosf(2*M_PI*RandUniform()));
 }
 
 // Returns poisson-distributed random number.  Uses Knuth's algorithm.
